rest of the site
The site was not working on the localhost and was not showing changes after they were made in the code so I chose to work off of the server.  I worked quite a bit off of the server and forgot that there was a github component to this and that is why there is a large gap between commits.